+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 
 # Version 1.0.5
 * Changed video to an animated gif.
+
+# Version 1.0.6
+* Added support for include and exclude options.
+* These can be set on a per project basis.